Moving DBIterator, DBRawIterator, Direction, and IteratorMode into separate module...
