+++ src/lib.rs
@@ -66,6 +66,7 @@ pub mod checkpoint;
 pub mod compaction_filter;
 mod comparator;
 mod db;
+mod db_iterator;
 mod db_options;
 mod db_vector;
 pub mod merge_operator;
@@ -73,9 +74,10 @@ mod slice_transform;
 
 pub use compaction_filter::Decision as CompactionDecision;
 pub use db::{
-    DBCompactionStyle, DBCompressionType, DBIterator, DBPinnableSlice, DBRawIterator,
-    DBRecoveryMode, Direction, IteratorMode, ReadOptions, Snapshot, WriteBatch,
+    DBCompactionStyle, DBCompressionType, DBPinnableSlice,
+    DBRecoveryMode, ReadOptions, Snapshot, WriteBatch,
 };
+pub use db_iterator::{DBIterator, DBRawIterator, Direction, IteratorMode};
 pub use db_vector::DBVector;
 
 pub use slice_transform::SliceTransform;
